+++ code/src/App.js
@@ -3,10 +3,8 @@ import GlobalFonts from './fonts/fonts'
 import { Provider } from 'react-redux'
 import { configureStore, combineReducers } from '@reduxjs/toolkit'
 import { Header } from './components/Header'
-import { AddTask } from './components/AddTask'
 import { TotalTasks } from './components/TotalTasks'
 import { Tasks } from './components/Tasks'
-import { ClearTasks } from './components/ClearTasks'
 import { todo } from 'reducers/todo'
 import { library } from '@fortawesome/fontawesome-svg-core'
 import { faTasks, faCheckCircle } from '@fortawesome/free-solid-svg-icons'
@@ -38,7 +36,7 @@ const AppContainer = styled.div`
     align-items: center;
     background-color: white;
     height: 80vh;
-    width: 500px;
+    width: 1108px;
     border-radius: 20px;
     -webkit-box-shadow: -4px 4px 41px -16px rgb(207, 207, 209);
     -moz-box-shadow: -4px 4px 41px -16px rgb(207, 207, 209);
@@ -54,9 +52,7 @@ export const App = () => {
           <AppContainer>
             <Header />
             <TotalTasks />
-            <AddTask />
             <Tasks />
-            <ClearTasks />
           </AppContainer>
         </Main>
       </Provider>
