+++ code/src/styles/Text.js
@@ -2,12 +2,13 @@ import styled from 'styled-components'
 
 export const Paragraph = styled.p`
   font-family: 'Avenir-Next-Regular';
-  color: ${(props) => props.color || 'black'};
+  color: ${(props) => props.color || 'darkgray'};
 `
 
 export const Title = styled.h2`
   font-family: 'Malina1';
   color: rgb(40, 95, 98);
+  font-size: 30px;
 `
 
 export const Subtitle = styled.h3`
